moved to mariaDb/mysql, added filter logic, added a simple command line tester
