+++ websmtp.sln
@@ -3,17 +3,16 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "websmtp", "websmtp\websmtp.csproj", "{E098E0BA-802A-4B26-B9D8-2B36BC62ED5D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "websmtp", "websmtp\websmtp.csproj", "{E098E0BA-802A-4B26-B9D8-2B36BC62ED5D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "tests", "tests\tests.csproj", "{46E73802-BD95-4F8D-8B67-7568DA7E4635}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "tests", "tests\tests.csproj", "{46E73802-BD95-4F8D-8B67-7568DA7E4635}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ManualTester", "ManualTester\ManualTester.csproj", "{4D545E43-22BA-474E-90B5-EE32A6835030}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{E098E0BA-802A-4B26-B9D8-2B36BC62ED5D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
@@ -24,5 +23,12 @@ Global
 		{46E73802-BD95-4F8D-8B67-7568DA7E4635}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{46E73802-BD95-4F8D-8B67-7568DA7E4635}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{46E73802-BD95-4F8D-8B67-7568DA7E4635}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4D545E43-22BA-474E-90B5-EE32A6835030}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4D545E43-22BA-474E-90B5-EE32A6835030}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4D545E43-22BA-474E-90B5-EE32A6835030}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4D545E43-22BA-474E-90B5-EE32A6835030}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
 	EndGlobalSection
 EndGlobal
