+++ websmtp/appSettings.json
@@ -5,7 +5,13 @@
       "Microsoft.AspNetCore": "Warning"
     }
   },
-  "Security":{
+  "Database": {
+    "Server": "cockpit.yvan.network",
+    "Name": "websmtp",
+    "Username": "websmtp",
+    "Password": "websmtp"
+  },
+  "Security": {
     "Username": "admin",
     "PasswordHash": "AQAAAAIAAAADAAAAEH4jJVZKzAbvotPnXpzZyFwJOxpWLkrvPu0HpEBQZ3A+bHThopydyArKxZsIC0CTZQ=="
   },
